moved utility function to util file and added tests
